WIP: This needs to be squashed, as it's not even in a runnable state.  I am switching branches.
